# Speed up empty highlighting many fields (backport of #61860)