… server/src/main/java/org/elasticsearch/search/fetch/subphase/highlight/UnifiedHighlighter.java